+++ src/Pages/Applications/DevtronApps/Details/AppConfigurations/AppConfig.tsx
