+++ src/components/app/details/main.tsx
@@ -30,7 +30,7 @@ import { MultiValue } from 'react-select'
 import { toast } from 'react-toastify'
 import { ErrorBoundary, sortOptionsByLabel } from '../../common'
 import { URLS } from '../../../config'
-import AppConfig from '../../../Pages/Applications/DevtronApps/Details/AppConfigurations/AppConfig1'
+import AppConfig from '../../../Pages/Applications/DevtronApps/Details/AppConfigurations/AppConfig'
 import { getAppMetaInfo } from '../service'
 import { AppMetaInfo } from '../types'
 import { EnvType } from '../../v2/appDetails/appDetails.type'
